Making the ball appear on the screen. Need to fic the ball into the paddle.

+++ Paddle.py
@@ -5,8 +5,8 @@ from Config import Config
 
 class Paddle:
 	def __init__(self, game_display):
-			self.x = (Config['game']['display_width'] * 0.45)
-			self.y = (Config['game']['display_height'] * 0.92)
+			self.x = int(Config['game']['display_width'] * 0.45)
+			self.y = int(Config['game']['display_height'] * 0.92)
 			self.game_display=game_display
 			self.paddle_width = Config['paddle']['paddle_width']
 			self.paddle_height = Config['paddle']['paddle_height']
